fix: Resolve all mypy type errors and add unified get_config method
- Fix 60 mypy type errors across 27 files
- Add get_config() method to ConfigurationProvider base class for unified interface
- Fix session_operations: store completed_at in metadata instead of non-existent attr
- Fix optigen_integration: use getattr for optional backend client methods
- Fix backend_session_manager: add type ignores for finalize_session return types
- Fix refactoring_utils: use new get_config() method
- Add type ignores for override method signatures, return types, and dynamic attrs
- Fix TypedDict bracket access vs dot notation in config_builder

+++ traigent/evaluators/base.py
@@ -56,8 +56,8 @@ except Exception:  # pragma: no cover - executed only when module missing
         llm: Any | None = None
         embeddings: Any | None = None
 
-    RagasConfig = _FallbackRagasConfig  # type: ignore[misc]
-    RagasConfigurationError = RuntimeError  # type: ignore[misc]
+    RagasConfig = _FallbackRagasConfig  # type: ignore[misc, assignment]
+    RagasConfigurationError = RuntimeError  # type: ignore[misc, assignment]
     compute_ragas_metrics = None  # type: ignore[assignment]
 
 
@@ -868,7 +868,7 @@ class BaseEvaluator(ABC):
                     if hasattr(r, "execution_time") and r.execution_time > 0
                 ]
                 if response_times:
-                    return sum(response_times) / len(response_times)
+                    return float(sum(response_times) / len(response_times))
 
         # Fallback: look for latency in context directly
         if "latency" in context:
